refactor(codegen): extract branches and latch into plugin folders
