+++ scripts/codegen/src/core/registry.ts
@@ -1,9 +1,11 @@
 import { a11yPlugin } from "../plugins/a11y/index.ts";
+import { branchesPlugin } from "../plugins/branches/index.ts";
 import { coveragePlugin } from "../plugins/coverage/index.ts";
 import { dependenciesPlugin } from "../plugins/dependencies/index.ts";
 import { examplesPlugin } from "../plugins/examples/index.ts";
 import { formControlPlugin } from "../plugins/formControl/index.ts";
 import { imperativePropsPlugin } from "../plugins/imperativeProps/index.ts";
+import { latchPlugin } from "../plugins/latch/index.ts";
 import { motionPlugin } from "../plugins/motion/index.ts";
 import { propsPlugin } from "../plugins/props/index.ts";
 import { tokensPlugin } from "../plugins/tokens/index.ts";
@@ -28,4 +30,6 @@ export const PLUGINS: readonly SubstratePlugin[] = [
   propsPlugin,
   formControlPlugin,
   imperativePropsPlugin,
+  branchesPlugin,
+  latchPlugin,
 ];
