+++ scripts/codegen/src/plugins/latch/index.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+import type { SubstratePlugin } from "../../core/plugin.ts";
+import { stateEntry } from "./schema.ts";
+
+export const latchPlugin: SubstratePlugin = {
+  name: "latch",
+  schema: {
+    atomic: { state: z.record(z.string().min(1), stateEntry).optional() },
+  },
+  emit: {},
+};
