+++ scripts/codegen/src/plugins/latch/schema.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export const stateEntry = z.strictObject({
+  type: z.literal("boolean"),
+  initial: z.boolean(),
+});
